Rename metropolitan_ervin_szabo_library package to ervinszabolibrary
