+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/ervinszabolibrary/ErvinSzaboLibraryBookCreator.java
@@ -1,4 +1,4 @@
-package com.reka.lakatos.searchbyisbn.crawler.metropolitan_ervin_szabo_library;
+package com.reka.lakatos.searchbyisbn.crawler.ervinszabolibrary;
 
 import com.reka.lakatos.searchbyisbn.document.Book;
 import com.reka.lakatos.searchbyisbn.document.CoverType;
@@ -9,7 +9,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 @Component
-public class BookCreator {
+public class ErvinSzaboLibraryBookCreator {
 
     private static final String ISBN13_REGEX = "((?:[\\dX]{13})|(?:[\\d\\-X]{17}))";
     private static final String ISBN10_REGEX = "((?:[\\dX]{10})|(?:[\\d\\-X]{13}))";
